\name{gfimmCPP-package}
\alias{gfimmCPP-package}
\alias{gfimmCPP}
\docType{package}
\title{
\packageTitle{gfimmCPP}
}
\description{
\packageDescription{gfimmCPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gfimmCPP}
\packageIndices{gfimmCPP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gfimmCPP}

Maintainer: \packageMaintainer{gfimmCPP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
